+++ jvm/src/main/scala/gpdviz/server/GpdvizServer.scala
@@ -4,9 +4,10 @@ import akka.actor.ActorSystem
 import akka.http.scaladsl.Http
 import akka.stream.ActorMaterializer
 import gpdviz.async._
+import gpdviz.config
 import gpdviz.config.configFile
 import gpdviz.config.cfg
-import gpdviz.data.{DbInterface, FileDb, MongoDb, PostgresDb}
+import gpdviz.data.{DbInterface, PostgresDb}
 
 import scala.concurrent.ExecutionContextExecutor
 import scala.io.StdIn
@@ -49,9 +50,7 @@ object GpdvizServer {
 }
 
 class GpdvizServer extends GpdvizService {
-  val db: DbInterface = cfg.mongo.map(new MongoDb(_))
-    .getOrElse(cfg.postgres.map(new PostgresDb(_))
-    .getOrElse(new FileDb("data")))
+  val db: DbInterface = new PostgresDb(config.tsConfig)
 
   implicit val system: ActorSystem = ActorSystem()
   implicit val executionContext: ExecutionContextExecutor = system.dispatcher